+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/multiplatform/targets/wasm/HubdleKotlinMultiplatformWAsmWAsiExtension.kt
@@ -1,17 +1,19 @@
-package com.javiersc.hubdle.project.extensions.kotlin.multiplatform.targets
+package com.javiersc.hubdle.project.extensions.kotlin.multiplatform.targets.wasm
 
 import com.javiersc.hubdle.project.extensions.HubdleDslMarker
 import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.project.extensions.kotlin.multiplatform.hubdleKotlinMultiplatform
 import com.javiersc.hubdle.project.extensions.kotlin.shared.HubdleKotlinMinimalSourceSetConfigurableExtension
 import javax.inject.Inject
+import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.provider.Property
 import org.gradle.kotlin.dsl.configure
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+import org.jetbrains.kotlin.gradle.targets.js.dsl.KotlinJsNodeDsl
 
 @HubdleDslMarker
-public open class HubdleKotlinMultiplatformWAsmJsExtension @Inject constructor(project: Project) :
+public open class HubdleKotlinMultiplatformWAsmWAsiExtension @Inject constructor(project: Project) :
     HubdleKotlinMinimalSourceSetConfigurableExtension(project) {
 
     override val project: Project
@@ -20,18 +22,24 @@ public open class HubdleKotlinMultiplatformWAsmJsExtension @Inject constructor(p
     override val isEnabled: Property<Boolean> = property { false }
 
     override val requiredExtensions: Set<HubdleEnableableExtension>
-        get() = setOf(hubdleKotlinMultiplatform)
+        get() = setOf(hubdleKotlinMultiplatform.wasm)
 
-    override val targetName: String = "wasm"
+    override val targetName: String = "wasmWAsi"
 
-    override fun Project.defaultConfiguration() {
+    @HubdleDslMarker
+    public fun nodejs(action: Action<KotlinJsNodeDsl> = Action {}) {
         lazyConfigurable {
             configure<KotlinMultiplatformExtension> {
-                wasmJs {
-                    browser()
-                    nodejs()
+                wasmWasi().apply {
+                    nodejs { //
+                        action.execute(this)
+                    }
                 }
             }
         }
+    }
+
+    override fun Project.defaultConfiguration() {
+        lazyConfigurable { configure<KotlinMultiplatformExtension> { wasmWasi() } }
     }
 }