+++ hubdle-gradle-plugin/testFunctional/resources/kotlin/multiplatform/publishing/normal-2/wasmWasi/main/kotlin/Platform.kt
@@ -1,0 +1,7 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+package com.kotlin.multiplatform.sandbox.project
+
+actual object Platform {
+    actual val name: String = "wasmWasi"
+}